% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCoordinate.R
\name{fetchCoordinate}
\alias{fetchCoordinate}
\title{fetchCoordinate}
\usage{
fetchCoordinate(address, n = 10)
}
\arguments{
\item{address}{The address}

\item{n}{The number of batch query, n = 10 by default}
}
\value{
a data.frame
}
\description{
get coordinate based on address
}
\details{
Title
}
