% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{geocoord}
\usage{
geocoord(address, n = 10)
}
\arguments{
\item{address}{The address}

\item{n}{The number of batch query, n = 10 by default}
}
\value{
a data.frame
}
\description{
get coordinate based on address
}
\details{
Title
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")
dgp <- function(n) {
  address <- data.frame(Number= 1:n,
                     address = c("北京大学", "清华大学", "武汉大学", "华中科技大学", "南京大学"))
return(address)
}

system.time( z1 <- geocoord(dgp(100)$address) )
system.time( z2 <- geocoord(dgp(1000)$address) )
}
