% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodist.R
\name{geodist}
\alias{geodist}
\title{Calculating the distance and travel time between origins and destinations}
\usage{
geodist(data, lon1, lat1, lon2, lat2, type)
}
\arguments{
\item{data}{the data set}

\item{lon1}{the longitude of the original point}

\item{lat1}{the latitude of the original point}

\item{lon2}{the longitude of the destination point}

\item{lat2}{the latitude of the destination point}

\item{type}{the travel methods (0 = direct distance, 2 = drive car; 3 = walk (only support the distance within 5km))}
}
\value{
a data frame including origin coordinates, destination coordinates, distance (meters) and travel time (seconds).
}
\description{
calculate the distance and travel time between origins and destinations
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxx")
x <- data.frame(
  a = c(104.0141, 104.0518, 104.0644, 104.0390, 104.1890, NA),
  b = c(30.66794, 30.64201, 30.64035, 30.66362, 30.65145, NA),
  c = c(104.0652, 104.0652, 104.0652, 104.0652, 104.0652, 104.0652),
  d = c(30.57851, 30.57851, 30.57851, 30.57851, 30.57851, 30.57851)
)
y <- geodist(data = x, lon1 = "a", lat1 = "b", lon2 = "c", lat2 = "d", type = 0)
z <- geodist(data = x, lon1 = "a", lat1 = "b", lon2 = "c", lat2 = "d", type = 1)
}
