% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocation.R
\name{geolocation}
\alias{geolocation}
\title{Convert the coordinates into formatted addresses}
\usage{
geolocation(data, longitude, latitude, ncore = 999)
}
\arguments{
\item{data}{The dataset, a data.frame or data.table}

\item{longitude}{The column having longitude}

\item{latitude}{The column having latitude}

\item{ncore}{The specific number of CPU cores used (ncore = 999 by default, which indicates maximum of CPU cores minus 1 )}
}
\value{
a data.table which adds the formatted address in the original data set.
}
\description{
Convert the coordinates into formatted addresses
}
\note{
The value of "longitude" or "latitude" should be digits in numeric or character format. If not, the function may return empty result for this coordinate automatically.
}
\examples{
\dontrun{
library(amapR)
options(amap.key = "xxxxxxxxxxxxxxxx")

# Completed data
test <- data.frame(n = 1:5000, lng = c(114.4345,104.0837), lat = c(30.51105, 30.63087))
results <- geolocation(data = test, longitude = "lng", latitude = "lat")

# When the column 'lng' has missing value
test <- data.frame(n = 1:5000, lng = c(114.4345,''), lat = c(30.51105, 30.63087))
results <- geolocation(data = test, longitude = "lng", latitude = "lat")

# When the column 'lng' has special characters
test <- data.frame(n = 1:5000, lng = c(114.4345,'?'), lat = c(30.51105, 30.63087))
results <- geolocation(data = test, longitude = "lng", latitude = "lat")
}

}
\references{
Amap. Official documents for developers: Web Service API. https://lbs.amap.com/api/webservice/summary
}
