% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomap.R
\name{geomap}
\alias{geomap}
\title{Getting the map data with sf format}
\usage{
geomap(adcode, level = "default")
}
\arguments{
\item{adcode}{The administrative code}

\item{level}{The level of map data, including province, city and district}
}
\value{
sf data
}
\description{
Getting the map data with sf format
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")

# China map
# (1) fetch and map at province level
china1 <- geomap(adcode = 100000, level = "province")
ggplot(china1) +
  geom_sf(fill = "white") +
  theme_bw()
# (2) fetch and map at city level
china2 <- geomap(adcode = 100000, level = "city")
ggplot(china2) +
  geom_sf(fill = "white") +
  theme_bw()
# (3) fetch and map at district level
china3 <- geomap(adcode = 100000, level = "district")
ggplot(china3) +
  geom_sf(fill = "white") +
  theme_bw()
# (4) fetch and map with missing value in the argment "level"
china4 <- geomap(adcode = 100000)
ggplot(china4) +
  geom_sf(fill = "white") +
  theme_bw()

# Province map
# (1) fetch and map at province level
sichuan1 <- geomap(adcode = 510000, level = "province")
ggplot(sichuan1) +
  geom_sf(fill = "white") +
  theme_bw()
# (2) fetch and map at city level
sichuan2 <- geomap(adcode = 510000, level = "city")
ggplot(sichuan2) +
  geom_sf(fill = "white") +
  theme_bw()
# Note: The argument of level is not allowed for the municipality or special administrative region, please use 'level = district'.
beijing1 <- geomap(adcode = 110000, level = "city")
beijing2 <- geomap(adcode = 110000, level = "district")
ggplot(beijing2) +
  geom_sf(fill = "white") +
  theme_bw()
# (3) fetch and map at district level
sichuan3 <- geomap(adcode = 510000, level = "district")
ggplot(sichuan3) +
  geom_sf(fill = "white") +
  theme_bw()
# (4) fetch and map with missing value in the argment "level"
sichuan4 <- geomap(adcode = 510000)
ggplot(sichuan4) +
  geom_sf(fill = "white") +
  theme_bw()

# City map
# (1) fetch and map at city level
chengdu1 <- geomap(adcode = 510100, level = "city")
ggplot(chengdu1) +
  geom_sf(fill = "white") +
  theme_bw()
# (2) fetch and map at district level
chengdu2 <- geomap(adcode = 510100, level = "district")
ggplot(chengdu2) +
  geom_sf(fill = "white") +
  theme_bw()
# (3) fetch and map with missing value in the argment "level"
chengdu3 <- geomap(adcode = 510100)
ggplot(chengdu3) +
  geom_sf(fill = "white") +
  theme_bw()
}
