% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymap.R
\name{mymap}
\alias{mymap}
\title{mymap}
\usage{
mymap(code, level = "default")
}
\arguments{
\item{code}{The administrative code}

\item{level}{The level of map data, including province, city and district}
}
\value{
sf data
}
\description{
Get the map data with sf format
}
\details{
Title
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")

# China map
# (1) map at province level
china1 <- mymap(code = 100000, level = "province")
ggplot(china1) +
 geom_sf(fill = "white") +
 theme_bw()
# (2) map at city level
china2 <- mymap(code = 100000,  level = "city")
ggplot(china2) +
 geom_sf(fill = "white") +
 theme_bw()
# (3) map at district level
china3 <- mymap(code = 100000,  level = "district")
ggplot(china3) +
  geom_sf(fill = "white") +
  theme_bw()
# (4) with missing value in the argment "level"
china4 <- mymap(code = 100000)
ggplot(china4) +
  geom_sf(fill = "white") +
  theme_bw()

# Province map
# (1) map at province level
sichuan1 <- mymap(code = 510000, level = "province")
ggplot(sichuan1) +
  geom_sf(fill = "white") +
  theme_bw()
# (2) map at city level
sichuan2 <- mymap(code = 510000, level = "city")
ggplot(sichuan2) +
  geom_sf(fill = "white") +
  theme_bw()
#Note: The argument of level is not allowed for the municipality or special administrative region, please use 'level = district'.
beijing1 <- mymap(code = 110000, level = "city")
beijing2 <- mymap(code = 110000, level = "district")
ggplot(beijing2) +
  geom_sf(fill = "white") +
  theme_bw()
# (3) map at district level
sichuan3 <- mymap(code = 510000, level = "district")
ggplot(sichuan3) +
  geom_sf(fill = "white") +
  theme_bw()
# (4) with missing value in the argment "level"
sichuan4 <- mymap(code = 510000)
ggplot(sichuan4) +
  geom_sf(fill = "white") +
  theme_bw()

# City map
# (1) map at city level
chengdu1 <- mymap(code = 510100, level = "city")
ggplot(chengdu1) +
  geom_sf(fill = "white") +
  theme_bw()
# (2) map at district level
chengdu2 <- mymap(code = 510100, level = "district")
ggplot(chengdu2) +
  geom_sf(fill = "white") +
  theme_bw()
# (3) with missing value in the argment "level"
chengdu3 <- mymap(code = 510100)
ggplot(chengdu3) +
  geom_sf(fill = "white") +
  theme_bw()




}
