% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorconvert.R
\name{transcoord}
\alias{transcoord}
\title{Transform the coordinates from other coordinate systems to Amap system}
\usage{
transcoord(data, longitude, latitude, coordsys = "autonavi", ncore = 999)
}
\arguments{
\item{data}{The dataset, a data.frame or data.table}

\item{longitude}{The column having longitude}

\item{latitude}{The column having latitude}

\item{coordsys}{The coordinate system of your original location data, such as "gps", "baidu", "mapbar" and "autonavi" (coordsys = "autonavi" by default)}

\item{ncore}{The specific number of CPU cores used (ncore = 999 by default, which indicates maximum of CPU cores minus 1 )}
}
\value{
a data.table which adds the transformed longitude and latitude using Amap System in the original data set
}
\description{
This function supports to transform the coordinates from three other coordinate systems (including baidu, GPS and mapbar) to Amap system
}
\note{
The value of "longitude" or "latitude" should be digits in numeric or character format. If not, the function may return empty result for this coordinate automatically
}
\examples{
\dontrun{
library(amapR)
options(amap.key = "xxxxxxxxxxxxxxxx")

# Completed data
test <- data.frame(n = 1:500, lng = c(114.4345,104.0837), lat = c(30.51105, 30.63087))
results <- transcoord(data = test, longitude = "lng", latitude = "lat", coordsys = "baidu")
}
}
\references{
Amap. Official documents for developers: Web Service API. https://lbs.amap.com/api/webservice/summary
}
