% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCoordinate.R
\name{fetchCoordinate}
\alias{fetchCoordinate}
\title{fetchCoordinate}
\usage{
fetchCoordinate(address)
}
\arguments{
\item{address}{The address}
}
\value{
a matrix
}
\description{
fetch coordinate based on address
}
\details{
Title
}
\examples{
library(gaodemap)
options(gaode.key = 'xxxxxxxxxxxxxxxx')

address = c('四川大学','北京大学','aaa',NA)

coordinate <-fetchCoordinate(address)

}
