% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCoordinate.R
\name{fetchCoordinate}
\alias{fetchCoordinate}
\title{fetchCoordinate}
\usage{
fetchCoordinate(data)
}
\arguments{
\item{data}{The address}
}
\value{
a dataframe
}
\description{
fetch coordinate based on address
}
\details{
Title
}
\examples{
testaddress = c('北京市朝阳区望京东路4号横店大厦','北京市海淀区上地信息路9号奎科科技大厦','aaa',NA)
coordinate <-fetchCoordinate(testaddress)

}
