% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCoordinate.R
\name{fetchCoordinate}
\alias{fetchCoordinate}
\title{fetchCoordinate}
\usage{
fetchCoordinate(address)
}
\arguments{
\item{address}{The address}
}
\value{
a data.frame
}
\description{
get coordinate based on address
}
\details{
Title
}
\examples{
library(gaodemap)

x <- data.frame(Number= 1:500,
                address = c("北京大学", "清华大学", "武汉大学", "华中科技大学", "南京大学", "中山大学", "四川大学", "中国科学技术大学", "哈尔滨工业大学", "复旦大学"))

system.time(z <- fetchCoordinate(x$address))

fetchCoordinate("四川大学")

}
