% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLocation.R
\name{fetchLocation}
\alias{fetchLocation}
\title{fetchLocation}
\usage{
fetchLocation(data, lon, lat)
}
\arguments{
\item{data}{name of the data, dataframe}

\item{lon}{the longtidute}

\item{lat}{the latitude}
}
\value{
a dataframe
}
\description{
fetch location based on longtitude and latitude
}
\details{
Title
}
\examples{
testdata = data.frame(
lat = c(39.934,40.013,40.047,NA,4444),
lon = c(116.329,116.495,116.313,NA,6666)
)
address <- fetchLocation(testdata, lon = 'lon', lat = 'lat')


}
