% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLocation.R
\name{fetchLocation}
\alias{fetchLocation}
\title{fetchLocation}
\usage{
fetchLocation(lon, lat)
}
\arguments{
\item{lon}{the longtidute}

\item{lat}{the latitude}
}
\value{
a dataframe
}
\description{
fetch location based on longtitude and latitude
}
\details{
Title
}
\examples{
library(gaodemap)
options(gaode.key = 'xxxxxxxxxxxxxxxx')

data = data.frame(
  lat = c(39.934,40.013,40.047,NA,4444),
    lon = c(116.329,116.495,116.313,NA,6666)
    )

 address <- fetchLocation(lon = data$lon, lat = data$lat)
 address <- fetchLocation(lon = 104.0665, lat = 30.57227)


}
