% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{Converting the addresses to coordinates}
\usage{
geocoord(address, n = 10)
}
\arguments{
\item{address}{The address}

\item{n}{The number of batch query, n = 10 by default}
}
\value{
a data.frame
}
\description{
Converting the address to coordinates
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")
geocoord("your address in Chinese format")
}
