% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{Converting the addresses to coordinates}
\usage{
geocoord(address)
}
\arguments{
\item{address}{The address}
}
\value{
data.table
}
\description{
Converting the address to coordinates
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")
geocoord("address in Chinese format")
}
