% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{Convert addresses into coordinates}
\usage{
geocoord(data, address, ncore = 999, nquery = 10)
}
\arguments{
\item{data}{The dataset, a data.frame or data.table}

\item{address}{The column name of address}

\item{ncore}{The specific number of CPU cores used (ncore = 999 by default, which indicates the maximum of CPU cores minus 1 were used in parallel computing if your CPU is less than 999 cores)}

\item{nquery}{The number of query per batch (nquery = 10 by default)}
}
\value{
a data.table which adds the formatted address, longitude and latitude in the original data set.
}
\description{
Convert addresses into coordinates
}
\note{
According to the official document of AMap Web Service API, the address in the data set should be in Chinese format.
If a address is in English or includes special characters (i.e., ?, -, >, _, etc.), the function may return empty result for this address automatically.
}
\examples{
\dontrun{
library(amapR)
options(amap.key = "xxxxxxxxxxxx")

# Note: The "address" is the column having Chinese addresses, and the data set named "test"
# should be a data.frame or a data.table.
results <- geocoord(data = test, address = "address")

# Set the specific number of CPU cores used
results <- geocoord(data = test, address = "address", ncore = 4)
}

}
\references{
Amap. Official documents for developers: Web Service API. https://lbs.amap.com/api/webservice/summary
}
