% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{Convert addresses into coordinates}
\usage{
geocoord(
  data,
  address,
  city = "",
  ncore = 999,
  nquery = 10,
  errormethod = "pass"
)
}
\arguments{
\item{data}{The dataset, a data.frame or data.table}

\item{address}{The column name of address}

\item{city}{Specify the city to query. This argument supports the city name in Chinese, the city name in pinyin, the administrative code of city or the city code defined by Amap.
By default, this argument is empty. For more information, see the Amap official documents at https://lbs.amap.com/api/webservice/guide/api/georegeo.}

\item{ncore}{The specific number of CPU cores used (ncore = 999 by default, which indicates the maximum of CPU cores minus 1 were used in parallel computing if your CPU is less than 999 cores)}

\item{nquery}{The number of query in each batch (nquery = 10 by default). This argument is used to avoid the http 413 error when the request url is too long.}

\item{errormethod}{It specifies how a task evaluation error should be handled. According to ".errorhandling" parameter in the foreach::foreach function, we use two methods for handling errors, including "remove" and "pass". If the value is "remove", the result for that task will not be returned, or passed to the .combine function. If it is "pass", then the error object generated by task evaluation will be included with the rest of the results. You may ask why we did not use the "stop" method. The reason is that it may generate frustrating results if we use "geocoord" function in a loop. The default value is "pass".}
}
\value{
a data.table which adds the formatted address, longitude and latitude in the original data set.
}
\description{
Convert addresses into coordinates
}
\note{
According to the official document of AMap Web Service API, the address in the data set should be in Chinese format.
If a address is in English or includes special characters (i.e., ?, -, >, _, etc.), the function may return empty result for this address automatically.
}
\examples{
\dontrun{
library(amapR)
options(amap.key = "xxxxxxxxxxxx")

# Note: The "address" is the column having Chinese addresses, and the data set named "test"
# should be a data.frame or a data.table.
results <- geocoord(data = test, address = "address")

# Set the specific number of CPU cores used and the number of query in each batch
results <- geocoord(data = test, address = "address", ncore = 4, nquery = 5)

# Specify the city to query
results <- geocoord(data = test, address = "address", city = "chengdu")
}

}
\references{
Amap. Official documents for developers: Web Service API. https://lbs.amap.com/api/webservice/summary
}
