% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{Convert addresses into coordinates}
\usage{
geocoord(data, address)
}
\arguments{
\item{data}{The dataset, a dataframe or data.table}

\item{address}{The column name of address}
}
\value{
data.table
}
\description{
Convert addresses into coordinates
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxx")
# addr is the column having addresses in the dataset named dat.
geocoord(data = dat, address = addr)
}
