% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoord.R
\name{geocoord}
\alias{geocoord}
\title{geocoord}
\usage{
geocoord(address, n = 10)
}
\arguments{
\item{address}{The address}

\item{n}{The number of batch query, n = 10 by default}
}
\value{
a data.frame
}
\description{
get coordinate based on address
}
\details{
Title
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxxxx")

}
