% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolocation.R
\name{geolocation}
\alias{geolocation}
\title{Converting the coordinates to addresses}
\usage{
geolocation(lon, lat)
}
\arguments{
\item{lon}{the longitude}

\item{lat}{the latitude}
}
\value{
a data frame
}
\description{
Converting the coordinates to addresses
}
\examples{
library(amap)
options(amap.key = "xxxxxxxxxxxxxxxx")

coordinate <- data.frame(
  lat = c(39.934, 40.013, 40.047, NA, 4444),
  lon = c(116.329, 116.495, 116.313, NA, 6666)
)

address <- geolocation(lon = coordinate$lon, lat = coordinate$lat)
address <- geolocation(lon = 104.0665, lat = 30.57227)
}
